changing to class
